Update approved-label.yml
This is my first contribution to the project. I’ve added the "pull_request: write" permission to the GitHub Actions workflow to ensure that the 'approved' label is correctly added to pull requests after review approval. It was a great learning experience, and I look forward to contributing more in the future